risk-scoring/requirements.txt
# AML Risk Engine 필수 라이브러리
# Python 3.10+

# 기본 데이터 처리
numpy>=1.26.0
pandas>=1.3.0

# API 서버
flask>=2.0.0
flask-cors>=3.0.0
flasgger>=0.9.7  # Swagger/OpenAPI 문서 생성

# 룰북 파싱
pyyaml>=6.0

# HTTP 요청
requests>=2.31.0  # OFAC SDN XML 다운로드용

# 유틸리티
tqdm>=4.65.0

# 그래프 분석 (MPOCryptoML 패턴 탐지용)
networkx>=3.3

# ============================================
# 머신러닝 (필수 - stage2_scorer.py에서 사용)
# ============================================
scikit-learn>=1.3.0
# 
# PyTorch (필요시)
# torch>=2.3.0
# torch-geometric>=2.5.3

